implement magnitude and normalize functions

+++ util.go
@@ -6,6 +6,6 @@ import (
 
 const floatEqualityThreshold = 1e-5
 
-func isFloatEqual(a, b float32) bool {
-	return math.Abs(float64(a)-float64(b)) <= floatEqualityThreshold
+func isFloatEqual(a, b float64) bool {
+	return math.Abs(a-b) <= floatEqualityThreshold
 }
